refactor(front): reorganizes components and adjusts imports (#998)
Closes: #998
