+++ front/src/lib/components/general/Loader.svelte
