+++ front/src/lib/components/general/NavToggle.svelte
