+++ front/src/routes/post/[slug]/+page.svelte
@@ -1,7 +1,7 @@
 <script>
   import ScrollTop from '$lib/layout/ScrollTop.svelte';
-  import Meta from '@components/Meta.svelte';
-  import PageTransition from '@components/PageTransition.svelte';
+  import Meta from '@components/general/Meta.svelte';
+  import PageTransition from '@components/general/PageTransition.svelte';
   import Post from '@components/posts/Post.svelte';
   import PostHero from '@components/posts/PostHero.svelte';
   import Flourish from '@layout/Flourish.svelte';
